[FEAT] 핀 상세 - 홈/포스트/해결하기

+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/pin/ApplyPinEmojiResponse.kt
@@ -1,5 +1,5 @@
 package com.issueissyu.fe.data.remote.dto.response.pin
 
 data class ApplyPinEmojiResponse(
-    val selectedEmojiId: Int? = null,
+    val selectedEmojiId: Long? = null,
 )
